feat: `chunk_buffer` config param
The remote file handlers will tend to buffer a given amount of data from the underlying stream.
For example, the FTP handler will buffer `chunk_size / 8192` (max 10 min 1) chunks read from the underlying stream.
The other handlers also buffer data.

The task is to make the size of the buffer configurable. Unit should be the number of FILE_CHUNK_SIZE chunks to buffer (so in the FTP example it would have a minimum of 80kn -> 80 chunks