chore: ignore commit
This is a squashed commit with outdated work that was scrapped. You can
safely ignore it as most of this stuff was removed.
The intent was to add a `AnyBackend` similar to `AnyConnection` etc, and
add a "sqlite" backend to be used in place of the mock backend.

This turned out to bring even more issues so the work was scrapped.
This commit also adds some doc items which describe what was the
intention and tradeoffs