fix: :art: improve backend `get_files` endpoint
Changed the backend's `get_file` endpoint to receive an optional query parameter with the location to get the files from. This way, instead of receiving a bunch of files that are inside folders, the endpoint can just return the immediate children of the selected path (or of root if no path was selected).