feat: ✨ SDK getters for buckets/files (#461)
* fix: :adhesive_bandage: fix upload receipt format + add range to downloads + remove package manager from package.json

* fix: :bug: add missing std feature gates to shc-common and shc-indexer-db

* fix: :art: improve backend `get_files` endpoint

Changed the backend's `get_file` endpoint to receive an optional query parameter with the location to get the files from. This way, instead of receiving a bunch of files that are inside folders, the endpoint can just return the immediate children of the selected path (or of root if no path was selected).

* feat: :sparkles: add getters for buckets and files to sdk

* fix: :art: run formatter

* fix: :rotating_light: fix linter issues

* style: :art: format package.json of the sdk e2e tests

* fix: :art: remove trailling comma

* chore: :package: update sdk pnpm lock

* docs: :memo: update SDK MSP client README

* feat: :sparkles: add added methods to msp client example

* style: :art: fix identation