feat: :sparkles: add migration for `msp` field in file system's `StorageRequestMetadata` (#628)
This PR adds the runtime migration that has to run when the next runtime
upgrade that includes #600 is executed.

This is required because in that PR we changed the
`StorageRequestMetadata` struct, removing the old `msp` field that was
an `Option<(ProviderIdFor<T>, bool)>` and adding a new, more
comprehensive field `msp_status` that's of type
`MspStorageRequestStatus`, and if we don't migrate storage when that
change gets included in the chain, the storage requests mapping won't be
able to be decoded properly and the runtime might panic or produce
undefined behaviour.
The idea (and why PR #600 was merged without this migration) was to
execute the migration when there were no storage requests open, but we
realized that since we don't have an effective way of ensuring this
(like a pause flag that disallows issuing storage requests, or something
like that) we ended going up with the safe route which is this
migration.

## ⚠️ Breaking Changes ⚠️

- **Short description**
  
The `StorageRequestMetadata` type has a new `msp_status` field of type
`MspStorageRequestStatus`, requiring a runtime migration of all open
storage requests.

- **Who is affected**
  
- Runtime managers of runtimes that use StorageHub since a runtime
upgrade + migration have to be run.

- **Suggested code changes**
  
  Add the migrations to the runtime:
  ```rust
  /// Migrations to run on runtime upgrade.
pub type Migrations =
(pallet_file_system::migrations::v1::MigrateV0ToV1<Runtime>,);
  
  ...

  /// Executive: handles dispatch to the various modules.
  pub type Executive = frame_executive::Executive<
    Runtime,
    Block,
    frame_system::ChainContext<Runtime>,
    Runtime,
    AllPalletsWithSystem,
    Migrations,
  >;
  ```

---------

Co-authored-by: Facundo Farall <37149322+ffarall@users.noreply.github.com>